+++ src/main/java/io/fabric8/maven/docker/access/UrlBuilder.java
@@ -101,11 +101,12 @@ public final class UrlBuilder {
             .build();
     }
 
-    public String pullImage(ImageName name, String registry) {
-        return u("images/create")
-                .p("fromImage", name.getNameWithoutTag(registry))
-                .p("tag", name.getDigest() != null ? name.getDigest() : name.getTag())
-                .build();
+    public String pullImage(CreateImageOptions options) {
+        final Builder urlBuilder = u("images/create");
+        if (options != null) {
+            urlBuilder.p(options.getOptions());
+        }
+        return urlBuilder.build();
     }
 
     public String pushImage(ImageName name, String registry) {